feat: options props 예시 추가
