+++ frontend/src/components/common/Select.tsx
@@ -1,6 +1,18 @@
 import { useEffect, useState, useRef } from "react";
 import Label from "./Label";
 
+/* options props 예시
+  const OPTIONS = [
+    { label: "라벨1", value: 1 },
+    { label: "라벨2", value: 2 },
+    { label: "라벨3", value: 3 },
+    { label: "라벨4", value: 4 },
+    { label: "라벨5", value: 5 },
+    { label: "라벨5", value: 6 },
+    { label: "라벨5", value: 7 },
+  ];
+*/
+
 export type OptionType = {
   label: string;
   value: any;
@@ -37,6 +49,8 @@ const Select = ({
     ? "select-container error"
     : "select-container";
 
+  const selectControlClassName = isModalOpen ? "select-control active" : "select-control";
+
   const onClickSelectControl = () => setIsModalOpen((prev) => !prev);
   const onClickSelectOutSide = (e: MouseEvent) => {
     const target = e.target as HTMLElement;
@@ -70,7 +84,7 @@ const Select = ({
       <Label id={id} isError={isError} isRequierd={isRequierd} label={label} />
       <div ref={selectRef} className="select" role="presentation">
         <div
-          className="select-control"
+          className={selectControlClassName}
           onClick={onClickSelectControl}
           role="presentation"
         >
